Add files via upload
Welcome to the personal portfolio project I created for my application for a Codsoft internship. This online application, which has a simple and responsive design, highlights my HTML and CSS abilities. An introduction, talents, projects, resume download, contact details, and a footer are all included in the portfolio. Please look around to learn more about my experience and areas of expertise.